Initial naive implementation of capability system
This commit implements the most basic version of the capability system,
without any restructuring of the code and without and detailed pass
throughs to invesitigate elements which may have been missed.
It also has not been tested for functionality.
This commit also adds a script to iteratively compile the app being
tested for each different possible capability level and output the
resulting code size.